[language-benches] Split out txn_bench into a separate binary.
Closes: #7519
