+++ language/benchmarks/benches/transaction_benches.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use criterion::{criterion_group, criterion_main, Criterion};
-use language_benchmarks::{move_vm::bench, transactions::TransactionBencher};
+use language_benchmarks::transactions::TransactionBencher;
 use language_e2e_tests::account_universe::P2PTransferGen;
 use proptest::prelude::*;
 
@@ -19,22 +19,4 @@ fn peer_to_peer(c: &mut Criterion) {
 
 criterion_group!(txn_benches, peer_to_peer);
 
-//
-// MoveVM benchmarks
-//
-
-fn arith(c: &mut Criterion) {
-    bench(c, "arith");
-}
-
-fn call(c: &mut Criterion) {
-    bench(c, "call");
-}
-
-fn natives(c: &mut Criterion) {
-    bench(c, "natives");
-}
-
-criterion_group!(vm_benches, arith, call, natives);
-
-criterion_main!(vm_benches);
+criterion_main!(txn_benches);
